job-hunting/job-hunting-agent/requirements.txt
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.0.0
google-genai>=1.0.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
